Added a client controller and changed most of the dependencies accordingly.
