+++ AutomobileRepairShop/Views/Shared/Appointments.cshtml
@@ -7,7 +7,7 @@
     <body>
         <div class="container-fluid">
             <div class="d-flex align-items-center h-custom-2 px-5 ms-xl-4 mt-5 pt-5 pt-xl-0 mt-xl-n5">
-                <form asp-controller="Home" asp-action="AddAppoint" method="post" style="width: 23rem;">
+                <form asp-controller="Client" asp-action="AddAppoint" method="post" style="width: 23rem;">
 
                     <h3 class="fw-normal mb-3 pb-3" style="letter-spacing: 1px;">Make an appointment</h3>
